requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.11.8
cffi==1.14.4
chardet==4.0.0
cloudinary==1.23.0
confusable-homoglyphs==3.2.0
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cors-headers==3.5.0
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-heroku==0.3.1
django-model-utils==4.1.0
django-registration==3.1.1
django-rest-knox==4.1.0
django-star-ratings==0.9.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==19.9.0
idna==2.10
importlib-metadata==1.7.0
oauthlib==3.1.0
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.2
whitenoise==5.2.0
zipp==3.4.0
